// src/lib/interfaces/contract-staffs/IContractRenewal.ts
export interface IContractRenewal {
    namaKakitangan: string;
    idSementara: string;
    noKadPengenalan: string;
    emel: string;
    category: string;
    applicationDate: string;
    hiredDate: string;
    contractExpiryDate: string;
    remainingDaysToContractEnd: string;
    appraisalSkillMark: string;
    appraisalInterpersonalMark: string;
    appraisalTotalMark: string;
    status: string;
}